Moved useWorkspace
